requirements.txt
asgiref==3.2.7
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2020.4.5.1
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
defer==1.0.6
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
evdev==0.7.0
flake8==3.8.2
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
louis==3.5.0
lutris==0.5.4
mccabe==0.6.1
netifaces==0.10.4
numpy==1.18.4
olefile==0.45.1
onboard==1.4.1
pandas==1.0.3
pexpect==4.2.1
Pillow==5.1.0
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pyflakes==2.2.0
pygobject==3.26.1
python-apt===1.6.5-ubuntu0.2-elementary14-ubuntu5.1.1
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
six==1.14.0
sqlparse==0.3.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
virtualenv==15.1.0
wget==3.2
xkit==0.0.0
xlrd==1.2.0
zipp==3.1.0
